feat: 네트워크 모듈

+++ Assets/Scripts/World/GridMeshBuilder.cs
@@ -20,7 +20,7 @@ public class GridMeshBuilder : MonoBehaviour
 
     int width;
     int height;
-    int[,] map = null;
+    byte[,] map = null;
 
     public void SetMap()
     {
